pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Deployment example"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tensorflow = "^2.20.0"
numpy = "^2.3.4"
fastapi = "^0.119.0"
pydantic = "^2.12.2"
uvicorn = "^0.37.0"
python-multipart = "^0.0.20"
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:start_server"